cambios en interfaces de erp y crm

+++ ERP/all_projects.php
@@ -91,7 +91,7 @@ if (isset($_POST['aprobar_cotizacion'])) {
             <div class="modal-body">
                 <p>Inicio de sesión exitoso.</p>
             </div>
-            <button class="close-btn" onclick="closeModal()">Cerrar</button>
+            <button class="close-btn btn-secondary" onclick="closeModal()">Cerrar</button>
         </div>
     </div>
 <?php endif; ?>
@@ -110,12 +110,14 @@ if (isset($_POST['aprobar_cotizacion'])) {
                 <option value="finalizado">Finalizados</option>
                 <option value="facturacion">Facturación</option>
             </select>
-            <a href="new_project.php" class="btn btn-success chompa">Nueva Cotización</a>
+            <a href="new_project.php" class="btn btn-info chompa">Nueva Cotización</a>
             <a href="ver_clientes.php" class="btn btn-info chompa">Clientes</a>
             <!--<a href="lista_cot.php" class="btn btn-info chompa">Cotizaciones</a>-->
-            <a href="delete_project.php" class="btn btn-danger chompa">Eliminar Proyecto</a>
+            <?php if ($username == 'admin'): ?>
+                <a href="delete_project.php" class="btn btn-danger chompa">Eliminar Proyecto</a>
+            <?php endif; ?>
             <form method="POST" action="">
-                <button type="submit" name="logout" class="btn btn-danger chompa">Cerrar sesión</button>
+                <button type="submit" name="logout" class="btn btn-secondary chompa">Cerrar sesión</button>
             </form>
         </div>
     </div>
